Refactor agent module into submodules and add explain command
- Split agent.rs into smaller focused modules under src/agents/
  - command_json.rs: CommandJSON struct and execution logic
  - semi_autonomous_command_line_agent.rs: Main agent implementation
  - traits.rs: Step and other traits
  - mod.rs: Module exports
- Add new explain command to arguments
- Update imports and references throughout codebase
- Maintain all existing functionality while improving organization
