+++ src/agents/mod.rs
@@ -1,0 +1,4 @@
+pub mod semi_autonomous_command_line_agent;
+pub mod command_line_explain_agent;
+pub mod command_json;
+pub mod traits;
